requirements.txt
altair==5.5.0
apportionment==1.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
branca==0.8.1
bs4==0.0.2
cachetools==5.5.1
charset-normalizer==3.4.1
click==8.1.8
et_xmlfile==2.0.0
folium==0.19.4
fuzzywuzzy==0.18.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.44.0
mapclassify==2.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpi4py==4.0.2
narwhals==1.25.2
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
protobuf==5.29.3
PuLP==3.0.0
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
pyogrio==0.10.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
shapely==2.0.7
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.42.0
streamlit_folium==0.24.0
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
xyzservices==2025.1.0
